First Commit of All Python Tasks & Projects
This is the first commit of all the Python tasks and projects I have managed to complete so far. This will accumulate with future commits for further Python tasks and projects that I will do in order to used as a reference for coding in more complex work that I will do as well as revision. In this way, I will hopefully be able to slowly grasp the understanding of core fundamentals by then to be able to do the complex work that comes.