Completed Programming Objects Programs
I have completed all the Programming Objects Programs and this will again help me in the future when it comes to any projects that involve number of classes to be implemented together. This section also allowed me to get the basic understand of the key concepts of OOP (Object Oriented Programming), which are Encapsulation, Polymorphism and Inheritance.