Started on Mad Libs Generator Project
I have started on the Mad Libs Generator Project, which is in 2 parts just like the Java coursework project I did for the drones, a console version and a GUI version. However, unlike with the drone project, the parts of this project are being done simultaneously to try to save a bit of time.