Space Invaders Coursework
This is the android studio coursework I ended up submitting in the end since the original idea, which was to be a 2d platform game with a movable character in a discoverable map, didn't work out. I followed an online tutorial which showed how to build the game, with the features to my game being changed. However, the product didn't work out as well as I had hoped it would. The game severely lags and the spaceship just simply drops for no reason and then the game gets stuck in a lag loop.