Game Music
This is the game music used in my level. However, the committed version will be in the wrong file type for the code, so anyone trying to download and test out this code will need to change the music file back to a .WAV file. This is due to the file size limit imposed by GitHub, which prevented me from committing files bigger than 25MB.