Python Programming Coursework Year 3
Here is the Python Programming Coursework that I did with another person in order to find out and understand about the influencers' use and spread on Twitter. For confidential reasons, the API keys and other key related personal info needed for this ipynb file to work are removed. Some of the python libraries used include Pandas, MatPlotLib and NetworkX. Pandas was used to create the pandas dataframes for each task whenever required. MatPlotLib was used to plot statistical graphs of any kind, such as bar graphs, line graphs and others. Finally, NetworkX was used in Task 3 only to plot out the networks of the influencer's influence over Twitter and calculate any statistical values related to networks.