Image Analysis Group Coursework Part 1
This is the first group coursework for Image Analysis that I did with 2 other people. In this coursework, we looked at improving the quality of the image by using a variety of methods, which included the fast-fourier transformation (FFT) and its filter, along with the sharpening filter and median filter.