HCI Group Coursework (FULL VERSION)
This is a group coursework in HTML, CSS and JS, which saw the group design and create a website surrounding COVID, the stats and how to manage the virus. This was created at the height of the pandemic and so was set as the relevant task at the time. This contains all the other parts of the site, as well as the quiz that I had created.