Updated Alarm Clock
Found a bug in the minutes section when user wants to set the alarm to go off within more than an 1 hour. Therefore I had to slightly modify the minutes_left variable by carrying out the modulus of the integral division of the original minutes_left value to correct the issue.