Updated ERS v4
Updated the ERS Program to v4.

Changes include updating the Tensorflow, Deepface and other relevant modules/packages, as the program v3 had stopped running due to outdated modules/packages. This has also led to some code changes as a result of the package/modules update.  For example, the way the cv2.putText was taking the emotional data from the Deepface analyse function initialised at the start of the program had to be modified and updated as the old method that was used in v3 was no longer working. Updated the credits page and added the QUIT/EXIT button on the menu window to allow the users to quit the program.